requirements.txt
fastapi
uvicorn
pydantic
sentence-transformers
faiss-cpu
numpy
scikit-learn
PyPDF2
requests
